refactor handlers into /game/handler
