+++ app/src/androidTest/java/com/github/fribourgsdp/radio/PlayerGameHandlerTest.kt
@@ -1,6 +1,7 @@
 package com.github.fribourgsdp.radio
 
 import android.view.View
+import com.github.fribourgsdp.radio.game.handler.PlayerGameHandler
 import com.github.fribourgsdp.radio.mockimplementations.FakeGameView
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.Timestamp
