+++ app/src/main/java/com/github/fribourgsdp/radio/game/GameActivity.kt
@@ -11,6 +11,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.Timer
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.handler.HostGameHandler
+import com.github.fribourgsdp.radio.game.handler.PlayerGameHandler
 import com.github.fribourgsdp.radio.game.prep.GAME_IS_HOST_KEY
 import com.github.fribourgsdp.radio.game.prep.GAME_KEY
 import com.github.fribourgsdp.radio.game.prep.GAME_UID_KEY
