+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/HostGameHandler.kt
@@ -1,6 +1,7 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.handler
 
 import android.util.Log
+import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.game.Game
 import com.github.fribourgsdp.radio.game.GameView
 import com.google.firebase.firestore.DocumentSnapshot
