+++ app/src/main/java/com/github/fribourgsdp/radio/game/handler/PlayerGameHandler.kt
@@ -1,5 +1,7 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.game.handler
 
+import com.github.fribourgsdp.radio.Database
+import com.github.fribourgsdp.radio.FirestoreDatabase
 import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.game.GameView
 import com.github.fribourgsdp.radio.util.NOT_THE_SAME
